+++ data/src/sdk/primarynetwork.ts
@@ -183,7 +183,7 @@ export class PrimaryNetwork extends ClientSDK {
    * List validators
    *
    * @remarks
-   * Lists details for validators. By default, returns details for all validators. Filterable by validator node ids and minimum delegation capacity.
+   * Lists details for validators. By default, returns details for all validators.  The nodeIds parameter supports substring matching. Filterable by validation status, delegation capacity, time remaining, fee percentage, uptime performance, and subnet id.
    */
   async listValidators(
     request: operations.ListValidatorsRequest,